fix tslint errors
