+++ packages/fast-development-site-react/src/components/site/index.tsx
@@ -10,7 +10,8 @@ import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { ellipsis, localizeSpacing, toPx } from "@microsoft/fast-jss-utilities";
 import ComponentWrapper from "./component-wrapper";
 import CategoryList from "./category-list";
-import SiteTitle, { SiteTitleBrand } from "./title";
+import SiteTitle from "./title";
+import SiteTitleBrand from "./title-brand";
 import SiteMenu from "./menu";
 import SiteMenuItem from "./menu-item";
 import SiteCategory from "./category";
